Add files via upload
In this Novel Project, I am addressing the challenge of predicting machine failures in the manufacturing industry. This domain involves the use of heavy machinery, which are susceptible to breakdowns and failures. A machine failure occurs when there is an abrupt and unforeseen decline in the performance of a machine. Due to the infrequent occurrence of failures, there is a significant issue of dealing with an imbalanced dataset. This necessitates specialized pre-processing techniques before applying Machine Learning algorithms. The impact of these failures on the industry is substantial, resulting in significant losses.

To tackle this problem, I plan to employ various machine learning methods, including Logistic Regression, K-Nearest Neighbors, Decision Tree classifier, and Support Vector Classifier. Additionally, I intend to explore the use of Neural Networks to enhance predictive accuracy. The key objective is to identify the most suitable predictive model based on the assessment of each classifier algorithm's efficiency. This assessment will be conducted using performance metrics such as accuracy, recall, and precision, among others. By meticulously evaluating the performance of different classifier algorithms, I aim to develop a robust predictive model that can help mitigate the impact of machine failures in the manufacturing industry.